+++ frontend/src/component/navBar.js
@@ -38,6 +38,11 @@ export default function renderNavBar() {
                           locale.navBar.rank
                       }</a>
                   </li>
+                  <li class="nav-item tp-nav-item">
+                      <a class="nav-link active" href="/dashboard">${
+                          locale.navBar.dashboard
+                      }</a>
+                  </li>
               </ul>
               <div class="btn-group" role="group" aria-label="Language">
                   <button id="languageEN" type="button" class="btn btn-secondary ${
